+++ include/libfly/minimise/LBFGS/lbfgs.hpp
@@ -14,6 +14,8 @@
 
 // You should have received a copy of the GNU General Public License along with openFLY. If not, see <https://www.gnu.org/licenses/>.
 
+#include <fmt/core.h>
+
 #include <cstddef>
 #include <memory>
 #include <optional>
@@ -77,7 +79,7 @@ namespace fly::minimise {
       /** @brief Print out debug info. */
       bool debug = false;
       /**
-       * @brief If provided each frame of the minimisation will be written to this file.
+       * @brief If provided at each frame of the minimisation ``out`` will be written to this file.
        *
        * It is the users responsibility to ensure the lifetime of ``fout`` is at least as long as the lifetime of the ``LBFGS`` object
        * and ensure only a single ``LBFGS`` object writes to ``fout`` at any one time. This really only exist for debugging purposes...
@@ -104,17 +106,119 @@ namespace fly::minimise {
      *
      * \endrst
      *
-     * @param out Final position of optimiser.
-     * @param in Initial position + per particle info.
+     * @param out Write final position/gradient here (forwarded to potential).
+     * @param in Inputs required by potential and at-least Position.
      * @param pot Potential to minimise.
-     * @param threads Number of openMP threads to use.
+     * @param num_threads Number of openMP threads to use.
      * @return true If a local minimum was found.
      * @return false If no local minimum was found.
      */
-    auto minimise(system::SoA<Position &> out,
-                  system::SoA<Position const &, TypeID const &, Frozen const &> in,
-                  potential::Generic &pot,
-                  int threads = 1) -> bool;
+    template <typename... P1, typename... P2>
+    auto minimise(system::SoA<P1...> &out, system::SoA<P2...> const &in, potential::Generic &pot, int num_threads = 1) -> bool {
+      // Check inputs
+
+      verify(in.size() == out.size(), "LBFGS minimizer inputs size mismatch, in={} out={}", in.size(), out.size());
+
+      // Clear history from previous runs;
+
+      m_core.clear();
+
+      double skin = std::max(std::pow(m_opt.skin_frac, 1. / 3.) - 1, 0.0) * pot.r_cut();
+
+      if (m_opt.debug) {
+        fmt::print("LBFGS: threads = {}\n", num_threads);
+        fmt::print("LBFGS: Skin = {}\n", skin);
+      }
+
+      // Avoid floating point comparison warnings.
+      constexpr std::equal_to<> eq;
+
+      if (double r_cut = pot.r_cut() + skin; !eq(std::exchange(m_r_cut, r_cut), r_cut) || !m_nl) {
+        m_nl = neigh::List(m_box, r_cut);
+
+        if (m_opt.debug) {
+          fmt::print("LBFGS: Reallocating neigh list\n");
+        }
+      }
+
+      out[r_] = in[r_];  // Initialise out = in;
+
+      m_nl->rebuild(out, num_threads);
+
+      pot.gradient(out, in, *m_nl, num_threads);
+
+      double trust = m_opt.min_trust;
+
+      double acc = 0;
+      double convex_count = 0;
+
+      for (int i = 0; i < m_opt.iter_max; ++i) {
+        //
+        double mag_g = gnorm_sq(out[g_]);
+
+        if (m_opt.debug) {
+          fmt::print("LBFGS: i={:<4} trust={:f} acc={:f} norm(g)={:e} rebuild={}\n", i, trust, acc, std::sqrt(mag_g), eq(acc, 0.0));
+        }
+
+        if (m_opt.fout) {
+          m_opt.fout->commit([&] {
+            (m_opt.fout->write(P1{}, out), ...);
+            ;  //< Write the positions of the atoms.
+          });
+        }
+
+        if (mag_g < m_opt.f2norm * m_opt.f2norm) {
+          return true;
+        } else if (convex_count >= m_opt.convex_max) {
+          if (m_opt.debug) {
+            fmt::print("LBFGS: Early exit - curvature\n");
+          }
+          return false;
+        }
+
+        auto &Hg = m_core.newton_step<Position, PotentialGradient>(out, out);
+
+        ASSERT(gdot(out[g_], Hg[del_]) > 0, "Ascent direction in lbfgs", 0);
+
+        // Limit step size.
+        Hg[del_] *= std::min(1.0, trust / gnorm(Hg[del_]));
+
+        // Add distance of most displaced atom
+        acc += std::sqrt((Hg[del_] * Hg[del_]).colwise().sum().maxCoeff());
+
+        // Update positions in real space
+        out[r_] -= Hg[del_];
+
+        if (acc > 0.5 * skin) {
+          m_nl->rebuild(out, num_threads);
+          acc = 0;
+        } else {
+          m_nl->update(Hg);
+        }
+
+        pot.gradient(out, in, *m_nl, num_threads);
+
+        pot.visit([&](auto const &underlying) {
+          if constexpr (std::is_same_v<remove_cref_t<decltype(underlying)>, potential::Dimer>) {
+            if (underlying.curv() > 0) {
+              convex_count += 1;
+            } else {
+              convex_count = 0;
+            }
+          }
+        });
+
+        double proj = gdot(out[g_], Hg[del_]);
+
+        if (proj < -m_opt.proj_tol) {
+          trust = std::max(m_opt.min_trust, m_opt.shrink_trust * trust);
+        } else if (proj > m_opt.proj_tol) {
+          trust = std::min(m_opt.max_trust, m_opt.grow_trust * trust);
+        }
+      }
+
+      return false;
+    }
 
   private:
     Options m_opt;
@@ -124,8 +228,6 @@ namespace fly::minimise {
     double m_r_cut = -1;
 
     std::optional<neigh::List> m_nl;
-
-    system::SoA<PotentialGradient> m_grad;
   };
 
 }  // namespace fly::minimise